Add Gaussian BN assets
