+++ causal-hub/src/assets/json/arth150.json
@@ -1,6 +1,6 @@
 {
   "name": "arth150",
-  "digraph": {
+  "graph": {
     "labels": ["4", "8", "13", "20", "26", "38", "47", "61", "63", "78", "81", "86", "93", "96", "100", "101", "108", "111", "126", "135", "144", "155", "161", "181", "187", "197", "198", "209", "211", "219", "226", "234", "245", "248", "256", "269", "272", "281", "289", "296", "299", "328", "331", "342", "360", "363", "368", "377", "378", "412", "414", "422", "443", "444", "452", "454", "460", "464", "479", "480", "496", "519", "537", "539", "540", "547", "554", "558", "560", "565", "570", "573", "576", "585", "596", "598", "600", "603", "622", "623", "627", "629", "636", "640", "651", "661", "665", "666", "677", "679", "686", "699", "712", "714", "726", "736", "738", "739", "758", "767", "778", "779", "781", "783", "786", "793", "798"],
     "edges": [
       ["539", "4"],
