+++ causal-hub/src/assets/json/ecoli70.json
@@ -1,6 +1,6 @@
 {
   "name": "ecoli70",
-  "digraph": {
+  "graph": {
     "labels": ["aceB", "asnA", "atpD", "atpG", "b1191", "b1583", "b1963", "cchB", "cspA", "cspG", "dnaG", "dnaJ", "dnaK", "eutG", "fixC", "flgD", "folK", "ftsJ", "gltA", "hupB", "ibpB", "icdA", "lacA", "lacY", "lacZ", "lpdA", "mopB", "nmpC", "nuoM", "pspA", "pspB", "sucA", "sucD", "tnaA", "yaeM", "yceP", "ycgX", "yecO", "yedE", "yfaD", "yfiA", "ygbD", "ygcE", "yhdM", "yheI", "yjbO"],
     "edges": [
       ["icdA", "aceB"],
