+++ causal-hub/src/assets/json/magic_irri.json
@@ -1,6 +1,6 @@
 {
   "name": "magic_irri",
-  "digraph": {
+  "graph": {
     "labels": ["HT", "FT", "AMY", "GL", "GW", "GTEMP", "G4432", "G2670", "G4744", "G3092", "G3486", "G1533", "G2639", "G3823", "G1378", "G3105", "G3102", "G3098", "G317", "G5167", "G1778", "G3872", "G4529", "G1440", "G5794", "G4668", "G2764", "G457", "G3862", "G3964", "G4109", "G2089", "G3219", "G3209", "G3222", "G3212", "G4573", "G1311", "G2949", "G6003", "G6010", "G6123", "G2815", "G3049", "YLD", "CHALK", "BROWN", "SUB", "G1888", "G1958", "G4156", "G4382", "G3136", "G4145", "G3106", "G1266", "G3927", "G5997", "G4553", "G2179", "G5006", "G3992", "G678", "G3925"],
     "edges": [
       ["FT", "HT"],
