+++ causal-hub/src/assets/json/magic_niab.json
@@ -1,6 +1,6 @@
 {
   "name": "magic_niab",
-  "digraph": {
+  "graph": {
     "labels": ["YR.GLASS", "HT", "YR.FIELD", "MIL", "FT", "G418", "G311", "G1217", "G800", "G866", "G795", "G2570", "G260", "G2920", "G832", "G1896", "G2953", "G266", "G847", "G942", "G200", "G257", "G2208", "G1373", "G599", "G261", "G383", "G1853", "G1033", "G1945", "G1338", "G1276", "G1263", "G1789", "G2318", "G1294", "G1800", "YLD", "FUS", "G1750", "G524", "G775", "G2835", "G43"],
     "edges": [
       ["MIL", "YR.GLASS"],
