+++ causal-hub/src/models/bayesian_network/gaussian/mod.rs
@@ -6,3 +6,6 @@ pub use potential::*;
 
 mod model;
 pub use model::*;
+
+mod sufficient_statistics;
+pub use sufficient_statistics::*;
