+++ causal-hub/src/models/continuous_time_bayesian_network/categorical/model.rs
@@ -196,21 +196,22 @@ impl CTBN for CatCTBN {
             "Graph labels and distributions labels must be the same."
         );
 
-        // Assert the labels of the parameters are the same as the graph parents.
-        assert!(
-            // Check if all vertices have the same labels as their parents.
-            graph.vertices().into_iter().all(|i| {
-                // Check if the labels of the parameters are in the parents.
-                graph
-                    .parents(&set![i])
-                    .into_iter()
-                    .eq(cims[i].conditioning_labels().iter().map(|j| {
-                        // Get the index of the label in the graph.
-                        graph.labels().get_index_of(j).unwrap()
-                    }))
-            }),
-            "Graph parents labels and conditioning labels must be the same."
-        );
+        // Check if all vertices have the same labels as their parents.
+        graph.vertices().iter().for_each(|&i| {
+            // Get the parents of the vertex.
+            let pa_i = graph.parents(&set![i]).into_iter();
+            let pa_i: &Labels = &pa_i.map(|j| labels[j].to_owned()).collect();
+            // Get the conditioning labels of the CIM.
+            let pa_j = cims[&labels[i]].conditioning_labels();
+            // Assert they are the same.
+            assert_eq!(
+                pa_i, pa_j,
+                "Graph parents labels and CIM conditioning labels must be the same:\n\
+                \t expected:    {:?} ,\n\
+                \t found:       {:?} .",
+                pa_i, pa_j
+            );
+        });
 
         // Initialize an empty graph for the uniform initial distribution.
         let initial_graph = DiGraph::empty(graph.labels());
